Source Salesforce: Convert to airbyte-lib (#33936)
