+++ airbyte-integrations/connectors/source-salesforce/setup.py
@@ -20,4 +20,9 @@ setup(
     extras_require={
         "tests": TEST_REQUIREMENTS,
     },
+    entry_points={
+        "console_scripts": [
+            "source-salesforce=source_salesforce.run:run",
+        ],
+    },
 )
